doom: fix C-return keybinding bug in org mode
I don't understand how this bug works and I'm only 90% sure the fix
works.

I introduced this bug a couple of commits ago. I made a copy-paste error
and as a result my doom config tried to clear a keybinding in the wrong
map: it tries to adjust the markdown keymap instead of the org keymap.

Somehow that causes the keybinding for `+default/newline-below` to take
precedence over `+org/insert-item-below` in org mode, even though the
former reports that it is in the global keymap when interrogated with
C-h k C-return.